Moved note table generation stuff into the counter object, so it should only be run once on program startup, instead of per note.
